requirements.txt
boto3>=1.34.0
PyYAML>=6.0
python-dotenv>=1.0.0
